previewFile on button click instead of onchange
Add a new "Preview" button.
call previewFile function onclick of the Preview Button,
instead of onchange of the file input box.

This allows users to reupload new images or change show name
without reloading the page (and doing a new search again).

We also need to keep a copy of the original JSON search result
so that we can repaint the original grid before randomly
inserting the new coverart.